mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.76.0",
    "mcp>=1.25.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
